Спринт №10 Добавлен список друзей и лайки - Внесены правки 2

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -12,7 +12,7 @@ import java.util.Set;
 @Builder
 public class Film {
 
-    private int id;
+    private long id;
     @NotBlank
     @NonNull
     private String name;
@@ -22,5 +22,5 @@ public class Film {
     private LocalDate releaseDate;
     @Positive
     private int duration;
-    private Set<Integer> likes;
+    private Set<Long> likes;
 }
